requirements.txt
pandas
psycopg[binary]
pyyaml